refactor: 요청 처리 로직을 Controller로 분리
