+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -2,14 +2,10 @@ package org.apache.coyote.http11;
 
 import java.io.IOException;
 import java.net.Socket;
-import java.util.HashMap;
-import nextstep.jwp.db.InMemoryUserRepository;
-import nextstep.jwp.exception.NotFoundException;
+import nextstep.jwp.controller.Controller;
 import nextstep.jwp.exception.UncheckedServletException;
 import org.apache.coyote.Processor;
-import org.apache.coyote.http11.common.ContentType;
 import org.apache.coyote.http11.common.HttpMethod;
-import org.apache.coyote.http11.common.StaticResource;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.slf4j.Logger;
@@ -47,36 +43,21 @@ public class Http11Processor implements Runnable, Processor {
     }
 
     private HttpResponse makeResponse(final HttpRequest httpRequest) throws IOException {
+        final var controller = new Controller();
         if (httpRequest.getMethod().equals(HttpMethod.GET)) {
-            return doGet(httpRequest);
+            return doGet(controller, httpRequest);
         }
         throw new UncheckedServletException("지원하지 않는 Http Method 입니다.");
     }
 
-    private HttpResponse doGet(final HttpRequest httpRequest) throws IOException {
+    private HttpResponse doGet(final Controller controller, final HttpRequest httpRequest) throws IOException {
         final var path = httpRequest.getPath();
         if (path.equals("/")) {
-            return HttpResponse.withStaticResource(new StaticResource("Hello world!", ContentType.TEXT_HTML));
+            return controller.showIndex();
         }
         if (path.equals("/login")) {
-            final var queryString = httpRequest.getQueryString();
-            final var queries = queryString.split("&");
-            final var h = new HashMap<String, String>();
-            for (var query : queries) {
-                final var split = query.split("=");
-                h.put(split[0], split[1]);
-            }
-
-            final var account = h.get("account");
-            final var user = InMemoryUserRepository.findByAccount(account)
-                    .orElseThrow(() -> new NotFoundException("존재하지 않는 사용자입니다."));
-            final var password = h.get("password");
-            if (user.checkPassword(password)) {
-                System.out.println("user = " + user);
-            }
-
-            return HttpResponse.withStaticResource(StaticResource.path("/login.html"));
+            return controller.showLogin(httpRequest);
         }
-        return HttpResponse.withStaticResource(StaticResource.path(path));
+        return controller.show(httpRequest);
     }
 }
